requirements.txt
amqp==5.2.0
asgiref==3.9.1
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
Django==5.2.4
django_celery_results==2.6.0
fonttools==4.58.5
html2text==2024.2.26
idna==3.10
kombu==5.3.5
markdown-it-py==3.0.0
mdurl==0.1.2
pillow==10.2.0
prompt_toolkit==3.0.51
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.17.2
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==6.2.0
requests==2.31.0
rich==13.7.1
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
